Move bg pixmap init back into shell.cpp, add Gui BgImage event for control, improve startup performance ~6x by assuming RGB32 return from XGetImage
